Moving and renaming classes, and removing redundant interfaces and such.  Also adding support for sending an arbitrary number of separate extra data types
