--- modules/adopter-registration-api/src/main/java/org/opencastproject/adopter/registration/IForm.java
@@ -1,35 +1,0 @@
-/*
- * Licensed to The Apereo Foundation under one or more contributor license
- * agreements. See the NOTICE file distributed with this work for additional
- * information regarding copyright ownership.
- *
- *
- * The Apereo Foundation licenses this file to you under the Educational
- * Community License, Version 2.0 (the "License"); you may not use this file
- * except in compliance with the License. You may obtain a copy of the License
- * at:
- *
- *   http://opensource.org/licenses/ecl2.txt
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
- * License for the specific language governing permissions and limitations under
- * the License.
- *
- */
-
-package org.opencastproject.adopter.registration;
-
-/**
- * Interface for the statistic registration form.
- */
-public interface IForm {
-
-  /**
-   * Overwrites the fields of this object with the data from another object of the same type.
-   * @param form The form with the data for merging.
-   */
-  void merge(IForm form);
-
-}
